Se agrega funcion GetLocation se obtiene posicion , se mejoran modelos

+++ internal/models/ship.go
@@ -2,6 +2,6 @@ package models
 
 type Ship struct {
 	Name     string   `json:"name"`
-	Distance float32  `json:"distance"`
+	Distance float64  `json:"distance"`
 	Message  []string `json:"message"`
 }
